requirements.txt
faiss-cpu>=1.11.0
langchain>=0.3.25
langchain-community>=0.3.23
langchain-core>=0.3.59
langchain-openai>=0.3.16
langchain-text-splitters>=0.3.8
loguru>=0.7.3
pandas>=2.2.3
pypandoc>=1.15
pypdf>=5.4.0
pypdf2>=3.0.1
pyyaml>=6.0.2
rich>=14.0.0
sentence-transformers>=4.1.0
streamlit>=1.45.0
tavily-python>=0.7.2
tqdm>=4.67.1